@@ -38,7 +38,7 @@
     "aesthetic-css": "^1.0.1",
     "bcrypt": "^5.0.1",
     "connect-ensure-login": "^0.1.1",
-    "connect-history-api-fallback": "^1.6.0",
+    "connect-history-api-fallback": "^2.0.0",
     "connect-mongodb-session": "^3.1.1",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
@@ -54,7 +54,7 @@
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
     "multer-gridfs-storage": "^5.0.2",
-    "node-graceful-shutdown": "1.1.0",
+    "node-graceful-shutdown": "^1.1.0",
     "nodemailer": "^6.7.2",
     "nunjucks": "^3.2.3",
     "oauth2orize": "^1.11.1",
@@ -70,7 +70,7 @@
     "passport-local": "^1.0.0",
     "request": "^2.88.2",
     "request-promise-native": "^1.0.9",
-    "rimraf": "^3.0.2",
+    "rimraf": "^5.0.1",
     "samlify": "^2.8.7",
     "socket.io": "^4.4.1",
     "textversionjs": "^1.1.3",
@@ -83,28 +83,28 @@
     "@emoji-mart/data": "^1.0.6",
     "@emoji-mart/react": "^1.0.1",
     "@picocss/pico": "^1.5.7",
-    "axios": "^0.27.2",
-    "babel-loader": "^8.2.4",
+    "axios": "^1.4.0",
+    "babel-loader": "^9.1.2",
     "classnames": "^2.3.1",
     "core-js": "^3.21.1",
     "css-loader": "^6.7.1",
-    "dompurify": "^2.3.6",
+    "dompurify": "^3.0.3",
     "emoji-mart": "^5.2.2",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.0",
     "immers-client": "^2.13.2",
     "mini-css-extract-plugin": "^2.6.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-intl": "^5.24.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-intl": "^6.4.2",
     "react-router": "^6.4.2",
     "react-router-dom": "^6.4.2",
     "sass": "^1.50.0",
-    "sass-loader": "^12.6.0",
+    "sass-loader": "^13.2.2",
     "standard": "^17.0.0-2",
     "style-loader": "^3.3.1",
     "webpack": "^5.72.0",
-    "webpack-cli": "^4.9.2",
+    "webpack-cli": "^5.1.1",
     "yargs": "^17.5.1"
   },
   "overrides": {
