@@ -1,8 +1,9 @@
-import React, { useContext, useEffect, useState } from 'react'
+import React, { useContext, useState } from 'react'
 import ServerDataContext from './ServerDataContext'
 import Layout from '../components/Layout'
 import Post from './Post'
 import { immersClient } from './utils/immersClient'
+import { useAsyncEffect } from './utils/useAsyncEffect'
 import { useParams } from 'react-router-dom'
 import LayoutLoader from '../components/LayoutLoader'
 
@@ -10,7 +11,7 @@ export default function Thread ({ taskbarButtons }) {
   const { activityId } = useParams()
   const { isInIframe } = useContext(ServerDataContext)
   const [activity, setActivity] = useState()
-  useEffect(async () => {
+  useAsyncEffect(async () => {
     const activitiyObj = await immersClient.activities.getObject(window.location.href)
     if (typeof activitiyObj.actor === 'string') {
       activitiyObj.actor = await immersClient.activities.getObject(activitiyObj.actor)
